Feature/exceptions

+++ src/main/java/com/example/vuelos/controllers/ClienteController.java
@@ -1,6 +1,8 @@
 package com.example.vuelos.controllers;
 
-import com.example.vuelos.dtos.ClienteDTO;
+import com.example.vuelos.controllers.dtos.ClienteDTO;
+import com.example.vuelos.controllers.dtos.ClienteRequestDTO;
+import com.example.vuelos.exceptions.ResourceNotFound;
 import com.example.vuelos.services.ClienteService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -24,27 +26,27 @@ public class ClienteController {
 
     @GetMapping("/{id}")
     public ResponseEntity<ClienteDTO> getClienteById(@PathVariable Long id) {
-        return clienteService.findById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+        return ResponseEntity.ok(clienteService.findById(id)
+                .orElseThrow(() -> new ResourceNotFound("Cliente no encontrado con id: " + id)));
     }
 
     @PostMapping
-    public ResponseEntity<ClienteDTO> createCliente(@RequestBody ClienteDTO clienteDTO) {
-        ClienteDTO createdCliente = clienteService.create(clienteDTO);
+    public ResponseEntity<ClienteDTO> createCliente(@RequestBody ClienteRequestDTO clienteRequestDTO) {
+        ClienteDTO createdCliente = clienteService.create(clienteRequestDTO);
         return ResponseEntity.ok(createdCliente);
     }
 
-
     @PutMapping("/{id}")
-    public ResponseEntity<ClienteDTO> updateCliente(@RequestBody ClienteDTO clienteToUpdate, @PathVariable Long id) {
-        return clienteService.update(id, clienteToUpdate)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<ClienteDTO> updateCliente(@RequestBody ClienteRequestDTO clienteRequestDTO, @PathVariable Long id) {
+        return ResponseEntity.ok(clienteService.update(id, clienteRequestDTO)
+                .orElseThrow(() -> new ResourceNotFound("Cliente no encontrado con id: " + id)));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCliente(@PathVariable Long id) {
+        if (!clienteService.findById(id).isPresent()) {
+            throw new ResourceNotFound("Cliente no encontrado con id: " + id);
+        }
         clienteService.delete(id);
         return ResponseEntity.noContent().build();
     }
