+++ src/main/java/com/example/vuelos/controllers/VueloController.java
@@ -1,6 +1,8 @@
 package com.example.vuelos.controllers;
 
-import com.example.vuelos.dtos.VueloDTO;
+import com.example.vuelos.controllers.dtos.VueloDTO;
+import com.example.vuelos.controllers.dtos.VueloRequestDTO;
+import com.example.vuelos.exceptions.ResourceNotFound;
 import com.example.vuelos.services.VueloService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,25 +25,26 @@ public class VueloController {
 
     @GetMapping("/{id}")
     public ResponseEntity<VueloDTO> getVueloById(@PathVariable Long id) {
-        return vueloService.findById(id)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+        return ResponseEntity.ok(vueloService.findById(id)
+                .orElseThrow(() -> new ResourceNotFound("Vuelo no encontrado con id: " + id)));
     }
 
     @PostMapping
-    public ResponseEntity<VueloDTO> createVuelo(@RequestBody VueloDTO vueloDTO) {
-        return ResponseEntity.ok(vueloService.create(vueloDTO));
+    public ResponseEntity<VueloDTO> createVuelo(@RequestBody VueloRequestDTO vueloRequestDTO) {
+        return ResponseEntity.ok(vueloService.create(vueloRequestDTO));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<VueloDTO> updateVuelo(@RequestBody VueloDTO vueloToUpdateDTO, @PathVariable Long id) {
-        return vueloService.update(id, vueloToUpdateDTO)
-                .map(ResponseEntity::ok)
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<VueloDTO> updateVuelo(@RequestBody VueloRequestDTO vueloRequestDTO, @PathVariable Long id) {
+        return ResponseEntity.ok(vueloService.update(id, vueloRequestDTO)
+                .orElseThrow(() -> new ResourceNotFound("Vuelo no encontrado con id: " + id)));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteVuelo(@PathVariable Long id) {
+        if (!vueloService.findById(id).isPresent()) {
+            throw new ResourceNotFound("Vuelo no encontrado con id: " + id);
+        }
         vueloService.delete(id);
         return ResponseEntity.noContent().build();
     }
